requirements.txt
google-genai
rich

